requirements.txt
flask
gunicorn
Jinja2
SQLAlchemy
python-dotenv